# dist/api.yml
openapi: 3.0.3
info:
  version: '1.0.0'
  title: 'Exercise 3.1'
  description: 'API - Web Development Environments'
...
paths:
  /Register:
    post:
      tags:
        - Auth
      summary: Add new user
      description: A new user from the client is added to the server
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "201":
          description: User created on site
        "400":
          description: Not all reqired argument was given!
        "409":
          description: Username is already taken
        "401":
          description: User already logged in
  /Login:
    post:
      tags:
        - Auth
      summary: Login to the website
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginScheme"
      responses:
        "200":
          description: >
            User was found in the DB and authenitcated successfully.
            The session ID is returned in a cookie named 'session'. 
            You should include this cookie in subsequent requests
          headers:
            Set-Cookie:
              schema:
                type: string
                example: session=87s97sdhgf279293f8w0e9
        "400":
          description: Not all reqired argument was given!
        "401":
          description: UserName or Password is incorrect || User already login
...
  /teams/teamFullDetails/{teamId}:
    get:
      summary: get full deatils about a team
      description: useful for team page
      tags:
        - Teams
      parameters:
        - in: path
          name: teamId
          description: team's id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: data about the team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "403":
          description: The team is not part of the Superliga
        "404":
          description: teamID is not exists
  /teams/getTeams:
    get:
      tags:
        - Teams
      summary: Return all the teams in the system
      responses:
        "200":
          description: data about all the teams in the system
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ObjectPreview"
        "204":
          description: There is no teams in the system
...
  /manage_league/getAllGames:
    get:
      tags:
        - Association Representative
      summary: Return all the games in the system
      responses:
        "200":
          description: data about all the games in the system
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GamePreview"
        "401":
          description: User is not authorized
  /manage_league/getReferees:
    get:
      tags:
        - Association Representative
      summary: Return all the referees in the system
      responses:
        "200":
          description: data about all the referees in the system
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ObjectPreview"
        "204":
          description: There is no referees in the system
  /manage_league/addGame:
    post:
      tags:
        - Association Representative
      summary: Add new game
      description: A new game is added to the server
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewGame"
      responses:
        "201":
          description: Game created on site
        "401":
          description: User is not authorized  
        "400":
          description: Not all reqired argument was given!
        "404":
          description: RefereeID is not exist || The game already accured  
  /manage_league/addResultToGame:
    put:
      tags:
      - Association Representative
      summary: Update result to a game
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/updateGameResult'
      responses:
        '200':
          description: The result of the game where update!
        '401':
          description: User is not authorized 
        '400':
          description: Not all reqired argument was given || Result of the game cant be lower than zero
        '404':
          description: The game has not yet been played || GameID is not exist         
  /manage_league/addListEventsToGame:
    post:
      tags:
      - Association Representative
      summary: Add list event to a game
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/updateListEvent'
      responses:
        '201':
          description: The list of events has been successfully updated to the game
        '400':
          description: Not all reqired argument was given || The parameters entered for the duration of the event description game are not good
        '401':
          description: User is not authorized 
        '404':
          description: The game has not yet been played || GameID is not exist
  /manage_league/addReferee:
    post:
      tags:
      - Association Representative
      summary: Add new referee to our system
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Referee'
      responses:
        '200':
          description: New referee added!
        '409':
          description: Username is already taken
        '400':
          description: Not all reqired argument was given!
        '401':
          description: User is not authorized   
  /manage_league/removeReferee/{refereeID}:
    delete:
      tags:
      -  Association Representative
      summary: Remove referee
      parameters: 
      - name: refereeID 
        in: path
        required: true
        schema:
          type: number 
      responses:
        '200':
          description: OK
        '401':
          description: User is not authorized  
        '404':
          description: The refereeID not found
...
components:
  schemas:
    User:
      type: object
      properties:      
        userName: 
          type: string
          minLength: 3
          maxLength: 8
          pattern: ^[a-zA-Z]+$
          description: The UserName must conatins 3 to 8 letters!
          example: Shahar

        firstName:
          type: string
          description: First name of the user
          example: Shahar
          
        lastName:
          type: string
          example: Tal
          description: Last name of ther user

        country:
          type: string
          description: Uses's country
          example: Israel
          
        password:
          type: string
          minLength: 5
          maxLength: 10
          format: password
          pattern: ^[a-zA-Z0-9]*$
          example: Shahar123
          
        email:
          type: string 
          format: email
          example: ShaharAP@post.bgu.ac.il
          description: The email of the user
          
        image:
          type: string
          format: URL
          description: The user profile image
    LoginScheme:
      type: object
      properties:
        userName:
          type: string
          minLength: 3
          maxLength: 8
          description: The UserName must conatins 3 to 8 letters!
          example: ShaharAP

        password:
          type: string
          minLength: 5
          maxLength: 10
          format: password
          description:  The Password must contains at least 5 to 8 characters, one letter and one digit
          example: Shahar123
...
    PlayerPreview:
      type: object
      properties:
        playerID:
          type: number
          description: The player ID
          example: 4
        
        name:
          type: string
          description: The player name
          
        image:
          type: string
          description: player's image
          format: URL
          example: https://cdn.sportmonks.com/images/soccer/players/11/84587.png
...
    GamePreview:
      type: object
      properties:
        gameID:
          type: number
          description: The gameID
          example: 5

        game_date:
          type: string
          description: Game date
          example: '2021-06-01'

        game_time:
          type: string
          description: Game time
          example: '16:30'

        home_team_ID:
          type: number
          description: The home team ID
          example: 5
          
        away_team_ID:
          type: number
          description: The away team ID
          example: 15

        referee_ID:
          type: number
          description: The referee ID
          example: 3

        field:
          type: string 
          description: The field of the game        
    NewGame:
      type: object
      properties:
        game_date:
          type: string
          description: game date
          example: "2021-06-01"

        game_time:
          type: string
          description: game date
          example: "16:30"   

        refereeID:
          type: number
          description: Thre RefereeID
          example: 3

        home_team_ID:
          type: number
          description: Home Team ID
          example: 85
          
        away_team_ID:
          type: number
          description: Home Team ID
          example: 939

        field:
          type: string 
          description: The field of the game
          example: Bloomfiled        
...
    eventType: 
      type: string 
      enum:
      - Goal
      - Offsife
      - Offense
      - Red Card
      - Yellow card
      - Injury
      - Subsitution
    AddEvent:
      type: object
      properties:
        event_time:
          type: string
          description: The time of the event
          example: '16:30'
        
        minute:
          type: number
          description: The minute where the event accured
          example: 20
          
        event_description:
          type: string 
          description: Description of the event
          example: first goal of the game
          
        type:
          $ref: '#/components/schemas/eventType'
          example: 'Goal | Offsife | Offense | RedCard | YellowCard | Injury | Subsitution'  
    Event:
      type: object
      properties: 
        eventID:
          type: number
          description: The event ID
          example: 1

        event_date:
          type: string
          description: The time of the event
          example: '01-01-2021'

        event_time:
          type: string
          description: The time of the event
          example: '16:30'
        
        minute:
          type: number
          description: The minute where the event accured
          example: 20
          
        event_description:
          type: string 
          description: Description of the event
          example: first goal of the game
          
        type:
          $ref: '#/components/schemas/eventType'
          example: 'Goal | Offsife | Offense | RedCard | YellowCard | Injury | Subsitution'
    Game:
      type: object
      required:
        - game_date
        - game_time
        - home_team
        - away_team
        - field
      properties:    
        gameID:
          type: number
          description: The game ID

        game_time:
          type: string
          description: game time

        game_date:
          type: string
          description: game date  

        home_result:
          type: number
          description: The result of the game for the home team
        
        away_result:
          type: number
          description: The result of the game for the away team  

        home_team_ID:
          type: number
          description: Home team ID
          
        away_team_ID:
          type: number
          description: Away team ID
          
        refereeID:
          type: number
          description: The game referee ID

        field:
          type: string 
          description: The field of the game
          
        event_list:
          type: array
          minLength: 3 
          items:
            $ref: '#/components/schemas/Event' 
...
    Team:
      type: object
      properties:
        notInFavorite:
          type: boolean
          description: Check if the team in the logged in user favorites
          example: true

        name:
          type: string
          description: team name
          example: København

        players:
          type: array
          items:
            type: object
            $ref: "#/components/schemas/PlayerPreview"

        future_games:
          type: array
          items:
            $ref: '#/components/schemas/Game'
        
        past_games:
          type: array
          items:
            $ref: '#/components/schemas/Game'    
        
        coachID:
          type: number
          description: The team coach ID
          example: 37571890

        coach:
          type: string
          description: The team coach name
          example: Alex Cohen 

        coach_image:
          type: string
          description: The team coach image
...
    updateGameResult:  
      type: object
      properties:
        gameID:
          type: number
          description: The game ID
          example: 1
        
        home_result:
          type: number
          description: The home team score
          example: 1  
      
        away_result:
          type: number
          description: The away team score
          example: 2
    updateListEvent:   
      type: object
      properties:
        gameID:
          type: number
          description: The game ID
          example: 1
          
        listEvents:
          type: array
          items:
            $ref: '#/components/schemas/AddEvent'
          description: 'Event type should be one of: Goal | Offsife | Offense | RedCard | YellowCard | Injury | Subsitution'             
...
    Referee:
      type: object
      properties:
        user:
          description: Referee details
          items:
            $ref: '#/components/schemas/User'

        training:
          type: string
          description: The training of the referee
          example: Superliga trainer
    ObjectPreview:
      type: object
      properties:
        ID:
          type: number
          description: Object ID
          example: 1

        name:
          type: string
          description: Object Name
          example: "Referee Name"
...